+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - added support for full width numbers in note names when linking notes with 
   unicode characters
   (for [Issue #144](https://github.com/pbek/QOwnNotes/issues/144))
+- added a workaround to keep the note count in the tag name and still being 
+  able to edit it without having to remove the note count from the name
+  (for [Issue #147](https://github.com/pbek/QOwnNotes/issues/147))
 
 ## 1.4.3
 - added **Japanese translation** (a big thank you to piyo!)